// src/app/core/services/model/planeacionProduccionDto.ts
/**
 * Mi API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PlaneacionProduccionDto { 
    idProducto?: string | null;
    nombreProducto?: string | null;
    idFamilia?: number;
    nombreFamilia?: string | null;
    semana?: string | null;
    cantidadesProducir?: number;
    estado?: boolean | null;
}